Added jacoco and cucumber reports to build
Jacoco plugin was added to master pom file. Jacoco coverage reports are generated
and later used in Hudson CI.

Cucumber reports are generated in test phase of build. Assembly of those reports was
added to assembly project. This assembly also has an index page that combines all
cucumber reporst on a single page. This page is published with Hudson Html publishing
plugin.

Signed-off-by: Uros Mesaric Kunst <uros.mesaric-kunst@comtrade.com>