Re-build docker file based on Fedora 25
The dockerfile was changes primarily to be able to run on OpenShift
as a batch job.

The reason to change it to Fedora was due to the fact that
java:8 images are rather unstable, dependencies where missing from
upstream compared to the provided image and thus the dockerbuild
actually re-installed java 7 during the installation of Maven.

Signed-off-by: Jens Reimann <jreimann@redhat.com>