Implement Kura style text area entries
Kura supports the use of text area field for entering multi line
text data. For this the "description" of a field is suffixed with an
information on how the field should be rendered.

This change implements this functionality inside Kapua, so that the
same entry fields which are available in Kura, are also available
in Kapua.

Signed-off-by: Jens Reimann <jreimann@redhat.com>