Replace Mortbay Jetty 6.x with a more recent version
This version replaces the GzipFilter from Mortbay Jetty 6.x with
a more recent version of Eclipse Jetty (9.1.x).

There is no CQ required for this change as this version of Jetty is an
Eclipse project.

Upgrading to any version after 9.1.x is not possible as this
functionality is actually deprecated starting with Jetty 9.2.0.

Signed-off-by: Jens Reimann <jreimann@redhat.com>