Shrink docker files and use non-root user
This change does switch to the openjdk:8u111-jre-alpine image in order
to reduce the size of the docker images significantly. The base image
is about 40MB where the Debian based image is about 120MB. It also uses
the JRE instead of the JDK (which is about 240MB).

All file system operations are also aggregated in order to save some
more space by collapsing multiple filesystem layers.

Additionally the entrypoints are not called by shell anymore in order
to pass through Unix signals properly.

Signed-off-by: Jens Reimann <jreimann@redhat.com>