Add a customizable footer snippet on the login page
This change does add a customizable HTMl snippet section in the login
page footer. It's content can be set using the Kapua settings for the
web console.

It is currently required for the Eclipse Foundation sandbox server,
showing some mandatory links like "Terms of service", …

But I guess it might come in handy for other use cases as well.

Signed-off-by: Jens Reimann <jreimann@redhat.com>