Fix a few things in the API doc
The described approach only works if you are working on the local
OpenShift machine. It doesn't work remotely.

This approach uses the OpenShift routing for accessing the services.

Also does the documentation make use of environment variables in order
to make it easier to copy and paste from the documentation to the
command line.

Signed-off-by: Jens Reimann <jreimann@redhat.com>