Keep a database instance open during the test
This change keeps a database connection open during the test, supporting
in-memory database scenarios.

Up until now this was not necessary as the liquibase client leaked
connection and this kept an instance open anyway. Leaking database
connections is not recommended for production setups so this was fixed
in eb9d18ee3886b9017acbed880d0c70ecd653e7f7, which now does break the
unit tests of they properly close the database connections.