Align all unit tests to use logback
Before this change part of the unit tests was using logback, part was
assuming to use logback but actually didn't and the other part was
using log4j.

As we need an extra CQ for slf4-logj4 and log4j 1.2.x is deprecated
anyway to logical choice is to go with logback.