Make Kapua re-buildable from scratch
Currently the REST API uses two swagger artifacts which are only
available from an external download site. However it should be
possible to re-build Kapua from scratch without the need to resort
to BLOB files which cannot be re-built.

This also speeds up the build process as it removes an additional
maven repository entry.