Switch to using "break" as sub-classes might now throw an exception
This closes #538. As long as the methods called in the switch cases
actually do throw an exception the code will work. However if someone
makes a sub-class and decides not to throw an exception, the next
case block will in fact the called which is most likely not the
expected behavior.