Fix a few bugs in the guice locator implementation
This change adds a few unit tests to check for a set of bugs in the
guice locator implementation. Like a class/factory not implementing
its interface, throwing the wrong exception type or error code.

It also provides a more testable version of the guice locator, which
allows specifying the resource to load and refrains from using
global variables when possible.