Extract auto-registration in dedicated module
This change does extract the auto-registration parts into a dedicated
module in order to clean up circular dependencies inside of Kapua.

Currently Kapua requires the use of "Domain" object instances which
are only defined in the "-internal" modules, which would result in
the "shiro" module to reference all other "-internal" modules in order
to set up the new account. However some other "-interal" modules again
reference on the "shiro" module, which results in circular dependencies.

The auto-registration functionality is now split up into an API and
a first "simple" provider module. This allows the "shiro" module to
load additional providers when possible without directly referencing
them.