Refactored MockedLocator with Guice proposal
Based on UserService unit tests, changes are introduced to MockedLocator.
MockedLocator can for now be used as it was and in addition also by using
Google Guice DI.

UserService tests ware changed to work with Guice. If this is correct it will
be refactored on all service unit tests.

As part of commit, feature file for User Service was cleaned of excessive tabs
and spaces. No content changes were made.

Signed-off-by: Uros Mesaric Kunst <uros.mesaric-kunst@comtrade.com>