Fix synch issue in unit tests
This change fixes a thread synchronization issue in the unit tests.
As the internal handleConnect/handleDisconnect calls are actually
passed on to the proxy instance, waiting for the future of the
first transport instance is not enough.

Now there is a specific listener which will actually wait for the
event to be delivered.