Clean up internals after refactoring
This change does clean up internals like method names, relocating
the module system to the SPI module as it is not really part of the API.
Also is the MQTT part of the module system now and the module context
properly allows to adapt to requested classes instead of requiring
casts. This allows to forward adapts to chained clients.