Added timeouts to embedded ES in MessageStoreServiceTest
As of previous commit time-outs are needed in slow environments for
embedded ES to start and stop. This fix might solve the issue on
Eclipse Hudson CI.
Locally runs also without this timeouts. If it shows successful it will
be provided as commandline parameter.

Included is also separate ES data directory per start of embedded engine.
As till now it was static defined per Embedded ES instance. This is not
very good for unit tests.

Signed-off-by: Uros Mesaric Kunst <uros.mesaric-kunst@comtrade.com>