QA steps moved to main from src - reuse of Cucumber steps
Cucumber steps for services are moved from test to main, making them
accessible in other modules for creating Cucumber tests.

org.eclipse.kapua:kapua-qa is maven dependency that includes this steps.

Refactored for usage with sql scripts that are part of QA module resources
in maven artifacts and not external files.

Signed-off-by: Uros Mesaric Kunst <uros.mesaric-kunst@comtrade.com>