QA steps moved to separate module qa-steps
Cucumber steps for services are moved from test to main, making them
accessible in other modules for creating Cucumber tests. kapua-qa
uses it as a dependency.

org.eclipse.kapua:kapua-qa-steps is maven dependency that includes this steps.

Refactored for usage with sql scripts that are part of QA module resources
in maven artifacts and not external files.

Signed-off-by: Uros Mesaric Kunst <uros.mesaric-kunst@comtrade.com>