First tests and impl of Account - User delete based on ServiceEvents
Simple Cuccumber scenario that creates account and user in that account.
Then created account is deleted, causing user to be deleted too.
This deletion was tested with foreign key connection between two entities.
Then FK was removed and test failed.

Implementation of service events was provided on Account delete and User
listening on delete event. User is deleted again.

More implementation is needed, but this is basic SE framework use.

Signed-off-by: Uros Mesaric Kunst <uros.mesaric-kunst@comtrade.com>