Fix for display of long user name
This solves issue wher use name is too long and is not correctyl displayed
in console header.
Name is shortende and ellipsis is added.
When mouse is hoovered over username its full name is displayed in tooltip.

Solution was first tried by setting style on button, but it didn't work, so
I had to resolve to changing console.css for button.

This solves issue 1191.

Signed-off-by: Uros Mesaric Kunst <uros.mesaric-kunst@comtrade.com>