Fixing multi select checkbox on selecting Targets for Job
Changes were needed on EntityGrid because implementation was not
according to GXT rules. Setting of grid was moved before onRender
method.

This was done in a manner that should leave existing usage of EntityGrid
in tact. Test of this fix should be extended to other girds not only
on this single one.

This fixes issue #1357

Signed-off-by: Uros Mesaric Kunst <uros.mesaric-kunst@comtrade.com>