Fixing wrong redraw on fast menu change
When moving through menu items with cursor this is very fast and
causes settings which is redrawn asynchronously to redraw when About is
already redrawn as next item. This causes display of Settings header in
footer over the copyright.

I added checking in Settings redraw if selection is still on settings and
if not redraw of settings is skipped.

This fixes issue #1254.

Signed-off-by: Uros Mesaric Kunst <uros.mesaric-kunst@comtrade.com>