Fix logout error is user refreshes browser
Error was caused by exception raised when user was looked up in
security filter. I moved user lookup latter in the code where it is
really needed.
This resolves this issue. If issue will reemerge it might be necessary to
move user lookup code into privileged block.

This fixes issue #1019

Signed-off-by: Uros Mesaric Kunst <uros.mesaric-kunst@comtrade.com>