Feature that checks size of browser window
This feature is added to console when console is created. Listener is
registered that reacts on browser window resizing.
Minimum window width and height are set and if window is too small UI
is blocked with dialog that informs user of that issue.

This could solve multiple issues that are based on UI being to small to
work correctly or issues when layout of UI changes when browser is resized
and is smaller than minSize set for UI components.

Min size is now set to 900 x 530 based on empirical UI usage observations.
It could be made even less if scroll bars would be introducet to tabbed
panels.

Signed-off-by: Uros Mesaric Kunst <uros.mesaric-kunst@comtrade.com>