Proposal for Entity Grid auto refresh
This is an example of automatic refresh of Entity grids.
Example is based on DeviceGrid, where it is useful to have dynamic
refresh of device status.

Implementation is based on client side timer that executes pull for
data and checks if data that is displayed in grid has changed. If data
changed it issues refresh of grid, if not it stays as it is. This prevents
unnecessary blinking of UI.

Data change is checked on check of optlock fields sum of displayed records.

Signed-off-by: Uros Mesaric Kunst <uros.mesaric-kunst@comtrade.com>