Solution to border issues (#1590)
* Solution to border issues

Solution is based on per view fixes to styling of borders. Fixes are not
very concise. Those fixes are almost hacks, but they solve the look and
feel of UI.

Solutions are based mainly on setStyleAttribute calls in separate views.
In some cases they are based on actually traversing DOM tree and fixing
attributes.

There might still be some problems which will be addressed in next commits.
There were namely many issues in this single issue and it is hard to fix
them with one go. Iterate!

This fixes issue #1536 

Signed-off-by: Uros Mesaric Kunst <uros.mesaric-kunst@comtrade.com>