Fixing button behaviour for webkit and gecko browsers - same look
Changed behaviour of buttons so that they behave same on webkit based and
mozilla browser.

Buttons in mozilla are changed so that they don't have border when they are in
focus. Also mozilla buttons are now blue when in focus, same as with chrome
and other webkit browsers.

Filter button is now not blue when not selected.

This fixes issue #1276

Signed-off-by: Uros Mesaric Kunst <uros.mesaric-kunst@comtrade.com>