Group field not validated correctly on edit (#1700)
This removes mandatory group field on edit dialog for device. This removes
condition that resulted in false negative validation of group field.
As group field is mandatory, that means that device can not be created
without group, it is at least No Group. That means that when user edits
device, it will always have group field set and it doesn't have to check
for empty field.
This should resolve this odd issue. I hope.

This solves issue #1631

Signed-off-by: Uros Mesaric Kunst <uros.mesaric-kunst@comtrade.com>