Long (integer) value validation in NumberField
While validating Long that is 64 bit values for integers there is wrong
validation of max value. Value is not validated correctly because
value is converted to Double before validating. This problem occurs because
Double can have mantisa which is 2^52 (signed). This brings validation to
that value.
To solve issue, max value is not set to Long.MAX_VALUE but to
4.503.599.627.370.496. If this is not good enough, NumberField provided
by GXT can not be used, as its validator uses Double for storing its values.

Other types should function correctly, as there is no limitation with conversion
to Double.

This solves issue #1634

Signed-off-by: Uros Mesaric Kunst <uros.mesaric-kunst@comtrade.com>