Hidden problem in NumberField
When adding steps to job, there is list of dynamic Job step properties.
In case that property is of type Long, NumberField is used for entry
field e.g. Timeout, display of that value didn't work, thou in DB it was
correct.
While inspecting code it might never worked at all.
I did change for Long, but other changes might be needed for other types,
if this was not related only to NumberField component.

This solves issue #1725

Signed-off-by: Uros Mesaric Kunst <uros.mesaric-kunst@comtrade.com>