Special characters problem in Jobs fix (#1762)
Solves issue: #1756
Added method to get unescaped value of JobStepName propery in GwtJobStep class, used in GwtJobStep update method. Made change in JobStepEditDialog, so that so that value of jobStepName is unescaped when populating the dialog. Updated JobTargetGrid so that the values on the grid and tooltips are unescaped as well.

Signed-off-by: ct-ajovanovic <aleksandra.jovanovic@comtrade.com>