Dialog window closes after invalid selection fix
Brief description of the PR.
Changed behaviour after invalid selection for some of the dialogs.

**Releted issue**
This PR fixes #1753

**Description of the solution adopted**
Made changes to the affected dialogs, so that they stay open, invalid
inputs selected and error messages are shown. Changes were made to the
KapuaExistingCredentialException class, so that the returned error
message and the constructor no longer contain the userId. This change
affected CredentialServiceImpl class, which was updated as well. Added
new SelectionChangedListeners for some of the comboBoxes, and updated
some of the existing listeners.

**Screenshots**
None

**Any side note on the changes made**
Handled a nullPoinerException when setting groupId in submit() method of
RolePermissionAddDialog class.

Signed-off-by: ct-ajovanovic <aleksandra.jovanovic@comtrade.com>