Test check of index name has been changed not to check status of index which is not important for these tests.
datastore.index.prefix system property is set to none in case it is not set
as it stays set from previous test. This causes tests to fail in case they are
run in succession.

Delays at startup of tests were increased just a bit. As there is not many tests
to multiply this it is OK.

Signed-off-by: Uros Mesaric Kunst <uros.mesaric-kunst@comtrade.com>