REST API tests in combination with service steps
REST API steps and service steps are combined in Gherkin scenarios.
This principle is used for test of issue 2011 where there are two
users in two different accounts, and user from one account can see
user in another account, using REST interface.

Mixing service steps and embedded jetty with rest steps is now working
and this is template for this kind of tests.

Signed-off-by: Uros Mesaric Kunst <uros.mesaric-kunst@comtrade.com>