Solving problem with javadoc:jar test steep
Moved javadoc:jar to last task on test stage of Travis.
I also added install to this step, so that artifacts are not downloaded.
I am not sure why artifacts from cache are not used.

Added clean to build, as cache is used and it should be cleaned between builds.

Signed-off-by: Uros Mesaric Kunst <uros.mesaric-kunst@comtrade.com>