Changed PR to behave as user was logged out
Change implies that user while deleted it is the same as it would be
logged out and not be able to login, as it doesn't exist any more.

No, new messages are required for that it is existing behaviour.

Added logout if deleted user tries to change password.
Added session logout, because it failed at next user login.

Possible fix for returning exception instead of null.

Signed-off-by: Ana Albic  <ana.albic@comtrade.com>