Display of modified and created by user fields
This is one way that would be possible to solve issue #2465.
Aleksandra tried with her PR #2534.

This approach is a bit more aggressive on user service, where it uses
KapuaId.ANY in query() methods to find all records.
Of course if permission check is executed before, user has to have permission
ANY to execute such a query.
In case of doPrivileged() this also shouldn't be a problem as it is
privileged after all.

Fixed created by / modified by on all grids and details.

Signed-off-by: Uros Mesaric Kunst <uros.mesaric-kunst@comtrade.com>