Upgraded docker-deploy.sh script, added REST API tests
I have added a "smart" timeout to the docker-deploy.sh script because there were some trobules wtih running Rest API tests.
The problem is, that once you run this script, docker containers get status "Running" although
they are not properly started just yet - they need a minute or so to properly start.
Without this addition, the Rest API tests will start as soon the docker-deploy.sh scripts completes and
the tests will fail, because API container is not deployed.
Therefore this timeout checks API container log for a certain string every 5 seconds (when this string
appears the container is up and running), to see if the container is properly running or not.
I have tested this and it works fine.

Second thing is added JSON file with REST API tests in /kapua/rest-api/tests/ folder.

Third thing is addition of "newman" CLI tool into the GHA .yml file, which actually runs the rest api
tests. This tool is available on GitHub Marketplace.

Fourth thing is added additional stage (rest api tests) into Travis.yml file, although Travis is becoming obsolete.

There are just a couple of tests added in the JSON file, becasue rest-api tests tend to get very long (a lot of lines) and I wanted to avoid CQ.
Signed-off-by: Leonardo Gaube <leonardo.gaube@endava.com>